refactor(setup): use 'data' mixin for properties
Using the 'beforeCreate()' hook to initialize the 'data' object is
clumsy and not idiomatic for a mixin-based plugin.

This patch instead creates a 'data' fragment dynamically, based on the
databases/selectors defined in the 'pouch' configuration option.

This fragment is then mixed in along with the rest of the API, which
creates the data properties on the Vue instances that use 'pouch'.